\name{Java.sparrow.notes}
\alias{Java.sparrow.notes}
\docType{data}
\title{Java sparrow note duration and frequency.}
\description{
This dataset reports the durations and frequencies of use for 22,970 notes in 676 undirected songs produced by 73 Java sparrows (Padda oryzivora). Songs were recorded by Masayo Soma and colleagues and notes were assigned to classes by Rebecca Lewis (Lewis et al. 2021; Lewis et al. 2023).}
\usage{data("Java.sparrow.notes")}
\format{
  A list with 22970 observations on the following 3 variables.
  \describe{
    \item{\code{duration}}{Amount of seconds the note was sung for}
    \item{\code{note}}{Identifier of note type}
    \item{\code{ID}}{Identifier of individual bird}
  }
}
\source{<https://figshare.manchester.ac.uk/articles/dataset/Code_and_Data_from_Lewis_et_al_2023_Animal_Behavior/22550149>}
\references{Lewis RN, Kwong A, Soma M, de Kort SR, Gilman RT (2023) Inheritance of temporal song features in Java sparrows. Animal Behaviour 206, 61-74}
\keyword{datasets}